requirements.txt
Appium-Python-Client==1.1.0
APScheduler==3.7.0
argcomplete==1.12.3
asyncpg==0.22.0
atomicwrites==1.4.0
attrs==20.3.0
brotlipy==0.7.0
case-repair-sdk==1.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==2.9.2
docopt==0.6.2
future==0.18.2
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
jsonschema==3.2.0
numpy==1.20.2
packaging==20.9
pip==21.0.1
pipreqs==0.4.10
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pycrypto==2.6.1
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.4
pytest-html==3.1.1
pytest-metadata==1.11.0
pytz==2021.1
requests==2.25.1
selenium==3.141.0
setuptools==52.0.0.post20210125
six==1.16.0
TestRunner==0.8.4
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==2.0.0
urllib3==1.26.4
websocket-client==0.59.0
wechat-at==2.1.11
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
yarg==0.1.9
zipp==3.4.1
